requirements.txt
alembic==1.7.6
certifi==2021.10.8
charset-normalizer==2.0.11
ciso8601==2.2.0
dill==0.3.4
ftx==1.2.0
gevent==21.12.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
loguru==0.6.0
Mako==1.1.6
MarkupSafe==2.0.1
multiprocess==0.70.12.2
numpy==1.22.2
pandas==1.4.0
pika==1.2.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
websocket-client==1.2.3
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
